+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - netstandard bumped to netstandard 2.1
 - MicroElements.Reflection.Sources replaced internal reflection stuff
 - PR#9 recursive ${include} from @i00lii
+- PlaceholdersConfigurationProvider now ignores case for property search 
 
 # 2.3.0
 - PlaceholdersConfigurationSource do not caches configuration (rebuilds on each build)